added debug toggle

+++ bot.py
@@ -4,21 +4,30 @@ import discord
 import datetime
 from dotenv import load_dotenv
 from discord.ext import commands
-import environ
-# import configparser
+
+# set to true to read bot info form the config file
+# or false to read from env variables
+debug = False
 
 load_dotenv()
 
-# config = configparser.ConfigParser()
-# config.read('config.ini')
 
-env = environ.Env(DEBUG=(bool, False))
-
-TOKEN = env('BOT_TOKEN')
-GUILD = env('GUILD')
-CHANNEL = env('CHANNEL')
-
-bot = commands.Bot(command_prefix='#')
+if debug == True:
+    import configparser
+    config = configparser.ConfigParser()
+    config.read('config.ini')
+    print(config.sections())
+    TOKEN = config['discord']['BOT_TOKEN']
+    GUILD = config['discord']['GUILD']
+    CHANNEL = config['discord']['CHANNEL']
+    bot = commands.Bot(command_prefix='*')
+else:
+    import environ
+    env = environ.Env(DEBUG=(bool, False))  
+    TOKEN = env('BOT_TOKEN')
+    GUILD = env('GUILD')
+    CHANNEL = env('CHANNEL')
+    bot = commands.Bot(command_prefix='#')
 
 @bot.event
 async def on_ready():
@@ -35,18 +44,39 @@ async def newdrink(ctx):
     channel = ctx.channel
     ingredient = []
 
-    def check(m):
-            return m.author == ctx.author and m.channel == channel
+    async def exit_check():
+        print(message_response.content)
+        if message_response.content == "exit":
+            try: 
+                await ctx.message.delete()
+            except: 
+                print("failed to delete ctx message")
+            try:
+                await send2.delete()
+            except:
+                None
+            await send.delete()
+            await message_response.delete()
+            return True
+        else:
+            return False
 
-    send = await ctx.send('Name of the drink: ')
+    def check(m):
+        return m.author == ctx.author and m.channel == channel
+
+    send = await ctx.send('if you want to quit at any time you can just say "exit"')
+    send2 = await ctx.send('Name of the drink: ')
     message_response = await bot.wait_for('message', check=check)
+    if await exit_check(): return
     name = str(message_response.content)
     await ctx.message.delete()
     await send.delete()
+    await send2.delete()
     await message_response.delete()
 
     send = await ctx.send('discription: ')
     message_response = await bot.wait_for('message', check=check)
+    if await exit_check(): return
     description = str(message_response.content)
     await send.delete()
     await message_response.delete()
@@ -54,6 +84,7 @@ async def newdrink(ctx):
     send = await ctx.send('image url: ')
     send2 = await ctx.send('imgur is a good place to upload!')
     message_response = await bot.wait_for('message', check=check)
+    if await exit_check(): return
     url = str(message_response.content)
     await send.delete()
     await send2.delete()
@@ -61,6 +92,7 @@ async def newdrink(ctx):
 
     send = await ctx.send('how many ingredients: ')
     message_response = await bot.wait_for('message', check=check)
+    if await exit_check(): return
     ingredients = int(message_response.content)
     await send.delete()
     await message_response.delete()
@@ -68,6 +100,7 @@ async def newdrink(ctx):
     for i in range(1,ingredients + 1):
         send = await ctx.send('ingredients #' + str(i) + ': ')
         message_response = await bot.wait_for('message', check=check)
+        if await exit_check(): return
         ingredient.append(str(message_response.content))
         await send.delete()
         await message_response.delete()
@@ -75,6 +108,7 @@ async def newdrink(ctx):
 
     send = await ctx.send('makeing instructions: ')
     message_response = await bot.wait_for('message', check=check)
+    if await exit_check(): return
     instructions = str(message_response.content)
     await send.delete()
     await message_response.delete()
@@ -91,5 +125,9 @@ async def newdrink(ctx):
     embed.add_field(name="instructions", value=instructions)
 
     await ctx.send(embed=embed)
+
+@bot.command()
+async def randomdrink(ctx):
+    print("random")
 
 bot.run(TOKEN)
